requirements.txt
absl-py==2.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.58.2
fsspec==2024.6.1
grpcio==1.73.0
Jinja2==3.1.4
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.3.0
pillow==11.0.0
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
setuptools==78.1.1
six==1.17.0
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
